Move math_utils header to the include folder and use better include structure
