+++ src/lib/local_lagrange/test/integration_test.cpp
@@ -10,8 +10,7 @@
 #include <utility>
 
 #include <boost/geometry/index/rtree.hpp>
-
-#include "../../math_utils/math_tools.h"
+#include <math_utils/math_tools.h>
 
 TEST(IntegrationTest, BuildAnLLF) {
