+++ src/lib/local_lagrange/test/lagrange_interpolant_test.cpp
@@ -1,4 +1,4 @@
-#include "../../math_utils/math_tools.h"
+#include <math_utils/math_tools.h>
 #include "../local_lagrange.h"
 #include "../local_lagrange_assembler.h"
 #include "../local_lagrange_interpolant.h"
