+++ src/lib/math_utils/CMakeLists.txt
@@ -1,6 +1,5 @@
 set(math_lib math_lib)
 set(math_src math_tools.cpp)
-set(math_hdr math_tools.h)
-add_library(${math_lib} ${math_src} ${math_hdr})
+add_library(${math_lib} ${math_src})
 #target_link_libraries( lagrange ${ARMADILLO_LIBRARIES})
 add_subdirectory(test)
