+++ src/lib/math_utils/test/math_test.cpp
@@ -1,4 +1,4 @@
-#include "../math_tools.h"
+#include <math_utils/math_tools.h>
 #include <gtest/gtest.h>
 #include <stdio.h>
